Prevent passUnwrappedFunctionsTransformer tests from forcing inlineWrappedFunctions tests to be run
