+++ test/helpers/transformCode.ts
@@ -1,0 +1,31 @@
+import ts from 'typescript';
+
+export function transformCode(
+  initialCode: string,
+  expectedCode: string,
+  transformer: ts.TransformerFactory<ts.SourceFile>
+): {
+  actual: string;
+  expected: string;
+} {
+  const source = ts.createSourceFile(
+    'elm.js',
+    initialCode,
+    ts.ScriptTarget.ES2018
+  );
+
+  const printer = ts.createPrinter();
+
+  const [output] = ts.transform(source, [transformer]).transformed;
+
+  const expectedOutput = printer.printFile(
+    ts.createSourceFile('elm.js', expectedCode, ts.ScriptTarget.ES2018)
+  );
+
+  const printedOutput = printer.printFile(output);
+
+  return {
+    actual: printedOutput,
+    expected: expectedOutput,
+  };
+}
