+++ test/inlineWrappedFunctions.test.ts
@@ -1,5 +1,6 @@
 import ts from 'typescript';
 
+import { transformCode } from './helpers/transformCode';
 import { createFunctionInlineTransformer } from '../src/transforms/inlineWrappedFunctions';
 
 test('it can process nested calls of A2 with non identifiers as the first arg ', () => {
@@ -201,32 +202,3 @@ test('it can inline functions declared not via an identifier or lambda', () => {
 
   expect(actual).toBe(expected);
 });
-export function transformCode(
-  initialCode: string,
-  expectedCode: string,
-  transformer: ts.TransformerFactory<ts.SourceFile>
-): {
-  actual: string;
-  expected: string;
-} {
-  const source = ts.createSourceFile(
-    'elm.js',
-    initialCode,
-    ts.ScriptTarget.ES2018
-  );
-
-  const printer = ts.createPrinter();
-
-  const [output] = ts.transform(source, [transformer]).transformed;
-
-  const expectedOutput = printer.printFile(
-    ts.createSourceFile('elm.js', expectedCode, ts.ScriptTarget.ES2018)
-  );
-
-  const printedOutput = printer.printFile(output);
-
-  return {
-    actual: printedOutput,
-    expected: expectedOutput,
-  };
-}
